final_project/project/requirements.txt
allure-pytest==2.8.13
allure-python-commons==2.8.13
apipkg==1.5
attrs==19.3.0
backcall==0.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
crayons==0.3.0
cryptography==2.9.2
decorator==4.4.2
execnet==1.7.1
Faker==4.0.3
Flask==1.1.2
idna==2.9
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py==1.8.1
pycparser==2.20
Pygments==2.6.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.2
pytest-forked==1.1.3
pytest-xdist==1.32.0
python-dateutil==2.8.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
SQLAlchemy==1.3.16
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webdriver-manager==2.4.0
Werkzeug==1.0.1
